Specify class name when storing Mongoid objects
Because of the namespacing, Mongoid attempted to store objects in different collections than we'd previously used. This meant it was unable to access data which had been added or seeded from another system. Specifying `store_in` fixes this issue.

The namespacing also interfered with Mongoid's ability to load embedded objects, so we add `class_name` to fix this.