Update README to cover engine
Our README should reflect that the repo is now for an engine rather than a standalone app. This commit adds some instructions for mounting it and removes some outdated info about running a Rails app.

We also remove the Hakiri badge as it's not used for gems. The front-office and back-office will be screened instead.