Sync dev and test secret keys across WCR service
In order for authentication to work across the apps the secret keys used need to be the same.

Currently the dummy testing app in this project is out of sync with the others so this change ensures the DEV and TEST secret keys are the same as in other projects.