Remove secret key validator as no longer used
Bascially in hindsight should have seen calling abort when the running in production and the secret key is a particular value would also cause the asset precompile to error.

The same thing is happening in that rails is being loaded by rake to carry out the precompile task, and needs to be able to complete loading without error for the task to complete.

Sticking abort() in blocks the task from completing.

However the good news is we have hit upon a 4th solution. There is logic that can be used to determine that

- we are running in production
- we were called from rake

Using this logic if we can deduce this particular scenario, we can programmtically set `config.secret_key` and appease Devise, allowing the precompile to complete.

So this change removes everything related to the validator, but leaves the logic to make available classes in the `lib/waste_carriers_engine` to the parent apps. Mainly because it was a slog to get it working, but also because there is value in having the option there in the future should we need it.