Add copy_cards field to registrations
Registrations from waste-carriers-frontend have an additional `copy_cards` field, as well as storing copy card info as OrderItems. For now, we just allow this field on the model so the two systems are compatible. We don't import it to the TransientRegistration as it's no longer relevant to this 'edition' of the registration.  However we will likely need additional changes in the future to make sure the behaviour of the field is the same as it is in the old system.