Cancelling Worldpay payment returns error message
https://eaflood.atlassian.net/browse/WC-394

Rather than showing a more informative message when cancelling a Worldpay payment, we get an 'Invalid response' error. This is caused because we are unable to build a valid MAC for a cancelled order if the user chooses to cancel their payment. In that case, the parameters we get back do not include the paymentAmount, the paymentCurrency or the paymentStatus. Instead we have orderAmount and orderCurrency.

We need to modify the params we receive for a cancelled payment so we can properly validate them, including the MAC.