Addresses don't allow firstName and lastName attributes
https://eaflood.atlassian.net/browse/WC-433

In waste-carriers-frontend, the firstName and lastName attributes are stored against the registration - however we also store them against the postal address. The engine doesn't allow this, which means it runs into issues when attempting to use addresses from the old system (for example, when updating a newly renewed registration).

We should allow these attributes to unblock renewals, but also make sure that we copy the firstName and lastName fields onto the postal address before renewing, so that we don't break any systems that rely on them (like mail merges).