Add names to contact address before renewing
We could add the names to the contact address when they are entered, or when the address is created – however this means we have to keep them in sync if the user goes back and modifies one or the other.

Since we don't use the names on the address at all during the renewal process, the safest option is just to copy them over as part of the renewal completion process.