Use ConvictionDataService to set up conviction info in DB
Previously we called the EntityMatchingService straight from the DeclarationFormController.

However, if a user declares they have a conviction, or if a match is found, we also want to create a new conviction_sign_off. Adding this to the EntityMatchingService would dilute its purpose too much. So instead we create a new service to call the EntityMatchingService and then set up data in the DB as required.