Use conviction_sign_offs to determine if check is pending
Previously we used a combination of conviction_search_results and the value of declared_convictions to determine whether a check was required and pending.

However, the conviction_sign_off object will only be created if a check is required, so now we can just use that.

This is more consistent with how the waste-carriers-frontend operates.