Add new method to build non-Worldpay payments
https://eaflood.atlassian.net/browse/WC-377

Initally we were building these payments in the back office: https://github.com/DEFRA/waste-carriers-back-office/pull/97 However, making creating a new payment the responsibility of the Payment model feels like a better fit. This also means we are more consistent with how we create different types of payments.

Also removed the manual_payment attribute, as after looking at the waste-carriers-service, it doesn't look like this is ever persisted or used for anything.