Add deprecated attributes to avoid DB errors
https://eaflood.atlassian.net/browse/WC-464

There are a number of attributes which were once in use in the existing system but have since been deprecated. However, some records in the database still have those fields, and attempting to import these records (eg to create a transient_registration) causes an error.

While the ideal solution would be to clean up our data, for now we should allow these attributes so we don't block users from renewing registrations which have them.