Take gem into account when building email_image_tag
This method was borrowed from waste-carriers-frontend. However, as that isn't an engine, the path it built for the engine didn't work when mounted into host applications (although tests did pass when run against the engine itself). waste-exemptions-shared turns out to be a better model to follow for working image attachments in an engine.