Send email when renewal application is received and pending
https://eaflood.atlassian.net/browse/WC-354

This task covers ensuring an email is sent when a user gets to the end of the renewal process but it is held as pending - this could be because of a pending payment or a conviction check. An email should be sent to confirm we have their details and the journey is complete.