Send different emails depending on payment status
If payment is pending, we need to send an email which includes instructions for payment. However, if the balance has been paid, we send a different email which just says the application has been received.

Content for these emails is still to be added/amended, but we can now determine which template to use depending on the balance.