Add relevant text to emails
The emails should now state what actions are required, if any, before the renewal is completed.

The pending payment one can reuse the text from the bank transfer page - this means that we only have to change that text in one location. The content on that page has also been through review so should be a better quality than the existing content.

We can remove @address_lines as it's not actually used for either renewal received email.