Changing from partnership to LLP blocks user journey on CH number page
https://eaflood.atlassian.net/browse/WC-451

Being able to register as an LLP is a new feature, so previously LLPs would have had to register as a limited company or a partnership. However, when changing the business_type during renewal from partnership to LLP, you cannot continue after entering the Companies House number for the LLP because it tells you that number has changed. This is because partnerships never enter a Companies House number, so we have nothing to compare it to.

This change disables the comparison if the business type for the registration used to be 'partnership'.