Store format of frontend dashboard link in I18n
This means the host application can override it with a different format if necessary, but the engine doesn't have to know anything about the host application for this to work (so no logic to check if the engine is currently mounted in the back office or not).

We can also reduce duplication by moving the dashboard_link method into the ApplicationHelper, instead of having it in two form objects.