Get different contact emails for digital and AD
We do this by checking if the current user's email matches the renewal account email. If they do, then the user who owns the registration is doing the renewal and it must be digital. If they don't, then another user is doing the renewal on their behalf and it must be assisted digital.