Add registration presenter (& presenters) to proj.
As we are just trying to replicate the existing PDF at this point, we are looking to migrate across the existing certificate template. However in review we found it relied on complex logic in the view, and lots of 'helper' methods which are actually on the model rather than as view helpers.

With all this logic needed to generate the view, I decided the cleanest solution was to introduce presenters.

> Presenters give you an object oriented way to approach view helpers.

I leaned heavily on http://nithinbekal.com/posts/rails-presenters/ and http://railscasts.com/episodes/287-presenters-from-scratch for the basis of my implementation. There is a gem called Draper (https://github.com/drapergem/draper) that allows you to decorate models with presentation logic, but I didn't feel this required adding a dependency, plus I preferred the idea of presenters as something completely separate from the model.

I currently have some initial spec's for the presenter, but I intend to clean/refactor them once I have settled on the final implementation for introducing PDF generation to the project.

How it works
------------

The `BasePresenter` inherits from SimpleDelegator, and this allows us with the line `super(model)` in the intializer to say any calls made on the presenter which are not recognised, pass down to the model passed in. Hence once `RegistrationPresenter` is instantiated you can call `@presenter.company_name` without having to explicitly expose a method or attribute on the `RegistrationPresenter`.

Often in presenters you still want access to the helpers available to the view (presenters sit between models and views). When you instantiate a presenter you need to pass in the `view_context` from the relevant controller or in our case mailer. It can then access any helper methods available in the view e.g. `link_to` by simply calling `h.link_to`. BTW creating a method called `h()` which exposes the view context appears to be a standard pattern for presenters.

With the `BasePresenter` in place `RegistrationPresenter` can focus on what specifically the view it is presenter for needs.