Updated version of `GeneratePdfService`
This is a cleaned up version of the GeneratePdfService, though I am torn over the need to have it.

As a separate service it means we can have a unit test that ensure at a basic level we can generate pdf's, without having to go through the mailer to generate one. However it doesn't do very much at the moment, and the line in the initializer could just as easily be copied into the mailer itself.

The alternative is to move the `render_to_string()` call currently sat in the renewal mailer into the service. However to do this the service then needs to know about ActionView and rendering, which already sits in the mailer.

So for now leaving as this till I have reached a conclusion.