Fix validation on copy card page
https://eaflood.atlassian.net/browse/WC-498

Our validation for the `temp_cards` field on the copy card form assumed that leaving it blank would sub in a zero, because the field is classed as an Integer in the database.

This is not the case - if the field already existed, the value is replaced with null, and if it doesn't exist then it's not created.

This means that when the order_items are created, checks against the value of the field fail and we get an error.

This commit modifies our validation to no longer allow blank values.