Deal with blank temp_card values when building Order
Order.new_order assumed that temp_cards will have a value. If it doesn't, then this causes the error which was initially reported.

Once all the renewals with blank temp_cards have passed through the system, this may not be necessary - so included a comment to remind us to review this in the future.