Add frontend cert., generate pdf & attach 2 email
In previous commits we have confirmed we can generate a pdf from the project, and we have introduced a presenter to handle any logic involved in what we display.

This change copies across the templates and layouts used by the frontend for generating a pdf certificate, altered to use the presenter, and stripping out any logic that had to handle lower tier registrations.

It also hooks up generating the pdf as part of sending the confirmation email.