Speed up email attachment tests
Prior to this change running just the attachment tests was taking more than 10 secs to run just 5 tests. My belief is this was because we were generating both an email and a certificate for each of the tests, when in fact the tests are just as valid using the same email for all of them.

I could have placed all the expectations into a single test, but then we lose the clarity. So instead I opted for a `before(:all)` block for this specific context. This runs just once, but each test can then access the variables it creates.

Now running these 5 tests takes just 3 seconds on my machine.