Ensure error in pdf does not block renewal
Though it won't be ideal for the user that they don't get their PDF, we don't want an error when trying to generate it and attach it to the email to block the renewal from completing. Hence we wrap the generation of the pdf in a rescue block, and return nil (and handle it) should any issues occur.