Enhance pdf generation error spec
One of the behaviours we expect should an error occur during pdf generation is that the resulting email will only have one attachment. I wanted to capture this in the spec, but again because the email tests are slow, I didn't want to create a separate test just for this.

I have tried moving all the setup into a `before(:all)` as previous, however this results in an error about mocks and stubs not being supported in that context.

So not knowing how this could be done, I've selected to just add the expectations to our existing test rather than adding another slow test.