Rename presenter based on view not model
When I built the initial instance of the presenter, I was following examples where the intention was to refactor view based logic currently based in a model out to its own domain.

However feedback in the PR made me rethink this intention. In my case I wanted to group all the logic for generating content in a particular view into a class I can more easily test, whilst avoiding adding lots of view helpers which would get loaded for all views.

So it makes sense to name our presenters based on the view they are bound to. This makes the scope and intention a lot more clearer.

Hence a chunk of renaming changes here, but hopefully the new name and folder structure makes things more clearer.