Remove duplication in reg. setup in tests
I had a couple of tests where I needed to have multiple main people, but there was nothing in the factories, so I hand cranked something in each test.

😳🤚💥

This refactors them to use a new trait I've added to the registration factory which creates multiple people in one place removing the duplication and hand cranking.