Add our own version of wicked_pdf_image_tag
Using Wicked PDF within an engine appears to mean that its helpers are not available to us when the engine is then mounted in a host app.

The one other reference to trying to do this suggests writing your own custom helper in place of using WickedPdf's (https://github.com/mileszs/wicked_pdf/issues/400).

Hence we add our own custom version of the helper, that points to a file within the engine.