Resolve confusion with WCRS_USE_XVFB_FOR_WICKEDPDF
When being review by another dev the assumption was that `WCRS_USE_XVFB_FOR_WICKEDPDF` shouldbe set to true in environments where no GUI exists, but set to false where there isn't one.

So when testing locally they were still getting errors because the script was being called, and failing because it could not find XVFB. And this was because they had set WCRS_USE_XVFB_FOR_WICKEDPDF to false.

As can be seen with the diff of this change, in fact the logic was simply does the env var exist. In retrospect (and foolish me for just copying and pasting) this assumption is valid and should be how we expect the app to behave.

Hence this change amends the logic to match how we would expect the app to behave when WCRS_USE_XVFB_FOR_WICKEDPDF is set to either true or false.