Allow renewal if the application was submitted but not immediate
Check if there is a matching TransientRegistration in the state of "renewal received". If there is, we know that their application to renew would not have been instant, as they only hit this page if there is a delay for payment or conviction checks. However, they must have met the other conditions for renewal when they began the process, as no TransientRegistration would have existed yet.

Note that we don't check for "renewal complete" as anything in this state would have been renewed immediately with no delay for NCCC checks.