Fix failing tests due to grace window changes
The grace window changes essentilly means simply marking a registration as `EXPIRED` no longer means it is expired.

So a number of tests that check you are redirected to a unrenewable page at various stages of the journey started failing because in the setup all they did was call `registration.metaData.expire!`.

This change also ensures that the expiry date is set to something that is outside the grace window, so our new code that determines whether something can be renewed returns false.