Rftr grace window checks due to naming confusion
In my rush to ship I added the check for whether a renewal is in the grace window to the `should_not_be_expired?` method. However as was rightly pointed our in the PR

> This method was initially intended to check whether the registration's status should be 'EXPIRED' and would be changed as soon as the scheduled job hit it, thus the concept of "should be expired".

If a registration was in the grace window, it should still be expired it just happens it can also be renewed.

Thankfully a simply fix, and much betterv place for the grace window check was to move it to the renewal_allowed? method.