Fix broken test caused by grace window
Found a test that was broken first by the order of checks in `renewal_allowed?`, but also when I started debugging found we should also be checking in `in_expiry_grace_window?` that the current day was not just before the last day of the grace window, but that it also came after the expiry date of the registration.

Also had a funny with the grace window being 0, so updated tests to account for this.