Fix conviction check to handle grace window
https://eaflood.atlassian.net/browse/WC-512

During testing of the grace window it was spotted that if a renewal was expired, and it matched during conviction checks, then it could not be continued because the button to access convictions approvals did not display.

This was because `TransientRegistration.pending_manual_conviction_check?` had a guard clause that checked the status was `ACTIVE`. This will no longer be the case for registrations which are EXPIRED but in the grace window, hence we change it to use `may_renew?`.

We also learnt doing this that `may_renew?` will only work if called on a `metaData` object that is linked to a registration, because of the call to `registration.identifier` in `CanChangeRegistrationStatus.renewal_application_submitted?`. Hence we had to tweak the tests and code to account for this.

Finally a test in the transient registration spec flagged that it was possible to for `may_renew?` to return true even if the registration in question was REVOKED. Hence a new guard in `CanChangeRegistrationStatus.renewal_allowed?` to check that the status of the registration is one that can be renewed.

🎉 WIP 🎉