Revert change to renewal_allowed? due to CC
Admittedly `renew_expired_registration?` was added to keep CodeClimate quiet, but the PR feedback rightly spotted that it wasn't returning a result in all cases.

From my read of the code the test that would cause us to miss all the previous logic gates in the method (and therefore return nil) was spec/models/waste_carriers_engine/registration_spec.rb:447 ("when the registration expiration date is less than 6 months away").

Debugging found this did cause nil to be returned, but the tests were still passing because nil == false i.e. we got the same result.

The issue was the name of the method did not really make it clear what it was doing, hence the logic inside didn't really make sense. After struggling to come up with a name and logic that made sense, and didn't break anything I instead came to the conclusion that CodeClimate was wrong in this case and leaving `renewal_allowed?` as it was before was the best solution.