Allow expired renewal to complete post declaration
https://eaflood.atlassian.net/browse/WC-526

We know now that we have renewals linked to registrations that have expired during the payment phase.

In a lot of these cases its because Worldpay has failed, and we have asked NCCC to use the workaround of sending the renewal back to the payment summary page and then completing it as a BAC's payment. The problem we are now encountering in greater numbers is renewals are expiring before the missed payment can be spotted and the workaround enacted.

We have therefore agreed that once a renewal has had the declaration accepted, it can still be renewed even if it then expires. This change covers the amendments necessary to support this.

---

🎉 WIP 🎉