Copy expiry date service from frontend
Adding the `ExpiryDateService` to the project as we only want to bother checking if the declaration has been confirmed in `CanChangeRegistrationStatus.renewal_allowed?` if the registration is expired. At the moment the check is just looking at the status, but we also need to check the date as well hence the need to bring this service across to this project.