Update ExpiryDateService to take reg. as arg
We also removes some methods that were needed on the frontend due to it holding values in milliseconds.

The reason we want to take a reg. is now that we have this service we can move methods off the CanChangeRegistrationStatus into this service, where from a design point of view I believe they are a better fit.