Fix failing BST to GMT test
With the changes in place we got one test failure which was

when the registration was created in BST and expires in GMT
  cannot be renewed when it reaches the expiry date plus 'grace window' in the UK

Essentially ExpiryCheckService.in_expiry_grace_window? was returning true and not false. This highlighted the importance of using consistent times and dates when handling GMT/BST hence the introduction of ExpiryCheckService.current_day.

I also rejigged how we setup the tests for BST/GMT related stuff to match what was done in the new ExpiryCheckService spec. This is more preference than requirement, but does mean we no longer double freeze time and ensure no possible issues due to lazy loading of the registration.