Add new renewal_declaration_confirmed? check
After some refactoring we now get to the meat & potatoes. We've implemented the check to see if the transient registration i.e. renewal has reached the declaration page and confirmed it. If this is the case then it does not matter of the registration has expired, the system will not block them from continuing.