Update reference to use search term
Realised that I hadn't duplicated all the filtering used in `TransientRegistrationFinderService` because I had overlooked the complexity of what we do with the search box.

Adding it as a named scope makes it more flexible and consistent with the other's we've added. However it does mean we have some duplication of the regex, and I don't know how to make both rubocop and Mongoid happy with the lambda.