Add missing in_progress search scope
Again another review of the existing search criteria highlighted that its not looking for those that are submitted, but using the inverse i.e. those where the workflow state is not one of the end pages.

Hence this adds the missing scope to match the "In progress" filter on the dashboard.