Add workflow states to conviction_sign_offs
https://eaflood.atlassian.net/browse/WC-483

The current implementation of the conviction checking workflow allows a flagged renewal to be checked, then approved or rejected. However, the actual process is more complex, and a renewal which does have matching convictions goes through an internal process before a final decision is made. This commit adds different states to the conviction_sign_off in order to better represent the workflow.