Implement AASM for conviction_sign_offs
Create a new concern to manage the conviction workflow stages and include it in the ConvictionSignOff model. A conviction_sign_off always starts as :possible_match. At this point it can be approved, rejected or move into :checks_in_progress when more work is required. Something in :checks_in_progress can then be approved or rejected.

We temporarily refer to the approve event as 'sign off' instead so as not to clash with the approve method currently on the ConvictionSignOff model.

These names and stages may change after further consultation with NCCC.