Update confirmed and confirmed_at as part of sign_off
Move the updates to these attributes out of the old approve method and into the new sign_off action.

This is part of getting rid of the old approve method entirely and relying on the state machine to handle all the work - moving confirmed_by will be slightly more complex, so we save that for the next commit.