Remove details in version page we do not use
The original design for the version page, which gives us the delivery team information about the particular app thats running assumed we would be updating the app version each time we released.

By the time we shipped, our release process dumped this aspect and instead just focuses on generating tags. So all we need to know is the commit hash, because then we can cross check that against the tags we have to confirm what version of a thing we have deployed.

Hence this change removes the app and engine versions from the version page.