Remove may_renew? from pending_manual_conviction_check?
A transient registration is pending a conviction check if

- its submitted
- the user declared a conviction, or the automated checker matched the organisation or an individual

Previously it was concerning itself with whether it could also be renewed. This is most likely due to the fact prior to the introduction of named mongoid scopes for searching, this would be used to filter out those that needed an approval to continue.

This change makes the decision that whether a registration can be renewed is irrelevant to the question are there outstanding conviction checks for this transient registration.

It also means one less hit on the db 😁