Refactor transient registration unit tests
Started by following http://www.betterspecs.org/#describe and ensuring where we are describing an instance method we start the description with a #.

However then to make this task a little easier, decided to reorder the tests slightly so that shared examples come first, then anything related to attributes, followed by the instance methods.

So did that, and the # in descriptions, but then spotted we were testing a method that actually came from a concern and was not directly on the `TransientRegistration` class.

This should really be in a shared example to make it clearer its not directly tied to the transient registration class, so also made that tweak.